Fix and Renable MAC and Windows Builds.  Needs more attention.
Had to diable building the RandomX Benchmark as it cannot find `version.h`
which is part of the Windows 8.1 SDK so build keeps failing in the pipeline
as it is using the depends cross compiler.

Builds fine on a Native Windows Machine so plan is to look into adding
the required SDK to depends.